typescript + start component for test
