+++ jsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": { 
+        "baseUrl": "src",
+        "strict": true,
+        "allowJs": true,
+        "jsx": "react-jsx"
+    },
+    "include": [
+        "src"
+    ]
+}
