+++ src/component/style/template.scss
@@ -1,0 +1,14 @@
+@import 'constant/style.scss';
+
+.convert_logo {
+	margin: 38px;
+}
+
+footer {
+	margin: 0 20px 15px 0;
+	div {
+		width: 468px;
+		padding: 12px 0 0 0;
+		background-color: $klein;
+	}
+}
