+++ src/constant/style.scss
@@ -1,5 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Lemonada:wght@300..700&family=Quicksand:wght@300..700&family=Raleway&display=swap');
-
+@import url('https://fonts.googleapis.com/css2?family=Lemonada:wght@300..700&family=Quicksand:wght@300..700&family=Rozha+One&display=swap');
 
 $off-white	 : #FEFEFE;
 $medium-grey : #999999;
@@ -10,4 +9,5 @@ $klein 		 : #002FA7;
 $light-orange: #E0A648;
 
 $lemonada : 'Lemonada', cursive;
+$RozhaOne : 'Rozha One', cursive;
 $quicksand : 'Quicksand', Roboto, Verdana, sans-serif;
