+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.scss';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import App from 'App.tsx';
+import reportWebVitals from 'reportWebVitals';
+import 'index.scss';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
