--- src/screen/HomeScreen.jsx
@@ -1,12 +1,0 @@
-import logo from '../logo.svg';
-
-const Home = () => {
-  return (
-    <header className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
-      <p>
-        Edit <code>src/App.js</code> and save to reload.
-      </p>
-    </header>)
-};
-export default Home;
